+++ src/api/call_fn.rs
@@ -190,8 +190,11 @@ impl Engine {
             let typ = self.map_type_name(result.type_name());
 
             result.try_cast().ok_or_else(|| {
-                let t = self.map_type_name(type_name::<T>()).into();
-                ERR::ErrorMismatchOutputType(t, typ.into(), Position::NONE).into()
+                let typename = match type_name::<T>() {
+                    typ @ _ if typ.contains("::") => self.map_type_name(typ),
+                    typ @ _ => typ,
+                };
+                ERR::ErrorMismatchOutputType(typename.into(), typ.into(), Position::NONE).into()
             })
         })
     }