+++ src/func/native.rs
@@ -321,8 +321,11 @@ impl<'a> NativeCallContext<'a> {
                 let typ = self.engine().map_type_name(result.type_name());
 
                 result.try_cast().ok_or_else(|| {
-                    let t = self.engine().map_type_name(type_name::<T>()).into();
-                    ERR::ErrorMismatchOutputType(t, typ.into(), Position::NONE).into()
+                    let typename = match type_name::<T>() {
+                        typ @ _ if typ.contains("::") => self.engine.map_type_name(typ),
+                        typ @ _ => typ,
+                    };
+                    ERR::ErrorMismatchOutputType(typename.into(), typ.into(), Position::NONE).into()
                 })
             })
     }
@@ -352,8 +355,11 @@ impl<'a> NativeCallContext<'a> {
                 let typ = self.engine().map_type_name(result.type_name());
 
                 result.try_cast().ok_or_else(|| {
-                    let t = self.engine().map_type_name(type_name::<T>()).into();
-                    ERR::ErrorMismatchOutputType(t, typ.into(), Position::NONE).into()
+                    let typename = match type_name::<T>() {
+                        typ @ _ if typ.contains("::") => self.engine.map_type_name(typ),
+                        typ @ _ => typ,
+                    };
+                    ERR::ErrorMismatchOutputType(typename.into(), typ.into(), Position::NONE).into()
                 })
             })
     }